Perform migrations only if required.
@pbadenski
@robyoung

+++ db/migrate/remove_department_from_policy.rb
@@ -4,7 +4,9 @@ require "dm-migrations/migration_runner"
 migration 2, :remove_department_from_policy do
   up do
     modify_table :policies do
-      drop_column :department
+      if adapter.field_exists?("policies", "department")
+        drop_column :department
+      end
     end
   end
 end
